requirements.txt
numpy==1.19.2
opencv_python==4.5.2.52
pyhocon==0.3.57
PyMCubes==0.1.2
tensorboard==2.6.0
torch==1.8.0
tqdm==4.50.2
trimesh==3.9.30
scipy==1.8.1
scikit-image==0.18.3
protobuf==3.20.*